+++ Code/LLDSA/Entities/FileOperator/LlsFileProcessorFactory.cs
@@ -1,0 +1,16 @@
+﻿using System.Dynamic;
+using LLDSA.Entities.FileOperator.LLSFileProcessor;
+
+namespace LLDSA.Entities.FileOperator
+{
+    public class LlsFileProcessorFactory
+    {
+
+        public static LlsFileProcessor GetLlsFileProcessor()
+        {
+            //Todo 
+            return null;
+        }
+
+    }
+}
