+++ Code/LLSDA.Tests/data/Shps/WorldCities/World_Cities.shp.xml
@@ -1,0 +1,3 @@
+<?xml version="1.0"?>
+<!--<!DOCTYPE metadata SYSTEM "http://www.esri.com/metadata/esriprof80.dtd">-->
+<metadata xml:lang="es"><Esri><MetaID>{2C500011-2F12-41E1-BFFE-61DC4C462661}</MetaID><CreaDate>20141229</CreaDate><CreaTime>15564400</CreaTime><SyncOnce>TRUE</SyncOnce><DataProperties><lineage><Process ToolSource="C:\Program Files (x86)\ArcGIS\ArcToolbox\Toolboxes\Data Management Tools.tbx\DefineProjection" Date="20141229" Time="155644">DefineProjection World_Cities GEOGCS['GCS_WGS_1984',DATUM['D_WGS_1984',SPHEROID['WGS_1984',6378137.0,298.257223563]],PRIMEM['Greenwich',0.0],UNIT['Degree',0.0174532925199433]] World_Cities</Process></lineage></DataProperties></Esri></metadata>
