+++ Code/LLSDA.Tests/data/Shps/WorldCities/World_Cities_README.txt
@@ -1,0 +1,4 @@
+HOW TO QUOTE THIS SHAPE:
+"World Cities". Downloaded from http://tapiquen-sig.jimdo.com. 
+Carlos Efraín Porto Tapiquén. Orogénesis Soluciones Geográficas. Porlamar, Venezuela 2015.
+Based on shapes from Enviromental Systems Research Institute (ESRI). Free Distribuition.
