+++ Code/LLSDA.Tests/data/Shps/WorldRivers/World_Hydrography_README.txt
@@ -1,0 +1,4 @@
+HOW TO QUOTE THIS SHAPE:
+"World Hydrography". Downloaded from http://tapiquen-sig.jimdo.com. 
+Carlos Efraín Porto Tapiquén. Orogénesis Soluciones Geográficas. Porlamar, Venezuela 2015.
+Based on shapes from Enviromental Systems Research Institute (ESRI). Free Distribuition.
